Negative reward for placing not stacking
